refactor(protocol): move Android TV codecs to src/backend/protocol/androidtv
Moves the three pure-codec files out of src/main/device/protocol/ into
src/backend/protocol/androidtv/, behind re-export shims at the old
locations (deleted in PR-3 when androidTvRemote.ts is itself broken up).

Files moved (git history preserved by --follow):
- src/main/device/protocol/certificate.ts
    -> src/backend/protocol/androidtv/certificate.ts
- src/main/device/protocol/pairingProtocol.ts
    -> src/backend/protocol/androidtv/pairing.ts
- src/main/device/protocol/remoteProtocol.ts
    -> src/backend/protocol/androidtv/remote.ts

The three old files become 1-line re-export shims so the existing
src/main/device/androidTvRemote.ts call site is unaffected.

This PR introduces the first Google TV non-regression gate described in
REFACTOR_PLAN.md §7.1 #8:

- 22 golden encode tests against captured hex fixtures
  (src/backend/protocol/androidtv/__fixtures__/remote.json) for setActive,
  pingResponse, every RemoteCommand keyInject, keyInjectRaw, voice
  begin/payload/end, and IME batch edits.
- 4 round-trip parse tests (encode then decode, assert shape).
- 2 framing invariants (length prefix, voice payload size).
- 6 certificate generator invariants (CN, self-signed, RSA-2048,
  long-lived validity window, unique serials).
- 1 PairingManager surface check (full FSM tests land in PR-5).

Test count: 35/35 passing. Coverage of the moved files is now substantial
even though we haven't turned on per-package thresholds yet.

Other small adjustments:
- tsconfig.backend.json: relaxed rootDir from src/backend to src so backend
  can import DTOs from src/shared, and added vitest/globals types so the
  __tests__/ folder typechecks under the same project.

Zero behavior change for the running app.
