+++ src/backend/protocol/androidtv/__fixtures__/remote.json
@@ -1,0 +1,24 @@
+{
+  "setActive_1": "0412020801",
+  "setActive_622": "05120308ee04",
+  "pingResponse_42": "044a02082a",
+  "keyInject_up": "06520408131003",
+  "keyInject_down": "06520408141003",
+  "keyInject_left": "06520408151003",
+  "keyInject_right": "06520408161003",
+  "keyInject_select": "06520408171003",
+  "keyInject_home": "06520408031003",
+  "keyInject_back": "06520408041003",
+  "keyInject_playPause": "06520408551003",
+  "keyInject_volumeUp": "06520408181003",
+  "keyInject_volumeDown": "06520408191003",
+  "keyInject_power": "065204081a1003",
+  "keyInject_assistantPress": "07520508e7011001",
+  "keyInject_assistantRelease": "07520508e7011002",
+  "keyInject_raw_powerLong": "065204081a1001",
+  "voiceBegin_session1": "05f201020801",
+  "voicePayload_session1_3bytes": "0afa010708011203010203",
+  "voiceEnd_session1": "058202020801",
+  "imeBatch_hello": "18aa0115080010001a0f0801120b080410041a0568656c6c6f",
+  "imeBatch_aim2_fc7": "14aa0111080210071a0b08011207080010001a0161"
+}
