+++ src/backend/protocol/androidtv/__tests__/pairing.test.ts
@@ -1,0 +1,18 @@
+import { describe, expect, it } from 'vitest';
+
+import { PairingManager } from '../pairing';
+
+/**
+ * PR-2 lands the module move + a light surface test. The full pairing FSM
+ * (happy path, invalid code, mid-flow disconnect, cert-rejected, restart-
+ * after-failure) is the subject of PR-5, which will inject a fake transport.
+ *
+ * Here we just assert the public class exists and can be constructed without
+ * exploding when given the inputs the real call site passes.
+ */
+describe('PairingManager — module surface', () => {
+  it('is exported as a class with the expected constructor surface', () => {
+    expect(typeof PairingManager).toBe('function');
+    expect(PairingManager.name).toBe('PairingManager');
+  });
+});
